baseline_model_pix2pix/requirements.txt

imageio>=2.9.0
matplotlib>=3.4.2

numpy>=1.20.3

pillow>=8.2.0

torch>=1.8.1
torchvision>=0.9.1

wandb>=0.10.33
